+++ src/content/components/DictPanelPortal/index.tsx
@@ -1,5 +1,7 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
+import memoizeOne from 'memoize-one'
+import { DictID } from '@/app-config'
 import CSSTransition from 'react-transition-group/CSSTransition'
 import DictPanel, { DictPanelDispatchers, DictPanelProps } from '../DictPanel'
 import { Omit } from '@/typings/helpers'
@@ -44,49 +46,65 @@ interface DictPanelState {
   readonly isDragging: boolean
 }
 
+const memoizeFrameHead = memoizeOne((selected: DictID[]): string => {
+  const meta = '<meta name="viewport" content="width=device-width, initial-scale=1">\n'
+
+  if (process.env.NODE_ENV === 'production') {
+    // load panel style and selected dict styles
+    return (
+      meta +
+      `<link type="text/css" rel="stylesheet" href="${browser.runtime.getURL('panel.css')}" />\n` +
+      selected.map(id =>
+        `<link rel='stylesheet' href=${browser.runtime.getURL(`/dicts/${isSaladictInternalPage ? 'internal/' : ''}${id}.css`)} />\n`
+      ).join('')
+    )
+  }
+
+  const styles = Array.from(document.querySelectorAll<HTMLLinkElement>('link[rel="stylesheet"]'))
+    .map(link => link.outerHTML)
+    .join('\n')
+
+  // remove wordEditor style
+  return meta + styles + `
+    <script>
+      document.querySelectorAll('link')
+        .forEach(link => {
+          return fetch(link.href)
+            .then(r => r.blob())
+            .then(b => {
+              var reader = new FileReader();
+              reader.onload = function() {
+                if (reader.result.indexOf('wordEditor') !== -1) {
+                  link.remove()
+                }
+              }
+              reader.readAsText(b)
+            })
+        })
+    </script>
+    `
+})
+
 export default class DictPanelPortal extends React.Component<DictPanelPortalProps, DictPanelState> {
   isMount = false
   root = isStandalonePage
     ? document.getElementById('frame-root') as HTMLDivElement
     : document.body
   el = document.createElement('div')
+  /** background layer when dragging to prevent event losing */
+  dragBg = document.createElement('div')
   frame: HTMLIFrameElement | null = null
   lastMouseX = 0
   lastMouseY = 0
 
-  frameHead = '<meta name="viewport" content="width=device-width, initial-scale=1">\n' + (
-    process.env.NODE_ENV === 'production'
-      ? `<link type="text/css" rel="stylesheet" href="${browser.runtime.getURL('panel.css')}" />`
-      : Array.from(document.querySelectorAll<HTMLLinkElement>('link[rel="stylesheet"]'))
-        .map(link => link.outerHTML)
-        .join('\n')
-        + `
-        <script>
-          document.querySelectorAll('link')
-            .forEach(link => {
-              return fetch(link.href)
-                .then(r => r.blob())
-                .then(b => {
-                  var reader = new FileReader();
-                  reader.onload = function() {
-                    if (reader.result.indexOf('wordEditor') !== -1) {
-                      link.remove()
-                    }
-                  }
-                  reader.readAsText(b)
-                })
-            })
-        </script>
-        `
-  )
-
-  state = {
+  state: DictPanelState = {
     isDragging: false,
   }
 
   constructor (props) {
     super(props)
     this.el.className = 'saladict-DIV'
+    this.dragBg.className = 'saladict-DragBg'
   }
 
   mountEL = () => {
@@ -140,6 +158,8 @@ export default class DictPanelPortal extends React.Component<DictPanelPortalProp
     if (this.frame) {
       this.frame.style.setProperty('will-change', 'top, left', 'important')
     }
+
+    document.body.appendChild(this.dragBg)
   }
 
   handleDragEnd = () => {
@@ -152,6 +172,8 @@ export default class DictPanelPortal extends React.Component<DictPanelPortalProp
     if (this.frame) {
       this.frame.style.removeProperty('will-change')
     }
+
+    this.dragBg.remove()
   }
 
   handleWindowMouseMove = (e: MouseEvent) => {
@@ -254,6 +276,7 @@ export default class DictPanelPortal extends React.Component<DictPanelPortalProp
   renderDictPanel = () => {
     const {
       isAnimation,
+      dictsConfig,
     } = this.props
 
     const {
@@ -279,7 +302,7 @@ export default class DictPanelPortal extends React.Component<DictPanelPortalProp
             bodyClassName='panel-FrameBody'
             name='saladict-dictpanel'
             frameBorder='0'
-            head={this.frameHead}
+            head={memoizeFrameHead(dictsConfig.selected)}
             frameDidMount={this.frameDidMount}
           >
             <DictPanel