+++ src/content/components/WordEditor/index.tsx
@@ -152,6 +152,10 @@ export class WordEditor extends React.PureComponent<WordEditorProps & { t: Trans
             />
             <label htmlFor='wordEditor-Note_Trans'>
               {t('wordEditorNoteTrans')}
+              <a
+                href='https://github.com/crimx/ext-saladict/issues/117'
+                target='_blank'
+              > Why?</a>
             </label>
             <textarea rows={5}
               name='trans'